test update/upgrade/resize/rotate

+++ magnum/drivers/common/templates/kubernetes/fragments/install-cri.sh
@@ -9,18 +9,25 @@ set -x
 
 ssh_cmd="ssh -F /srv/magnum/.ssh/config root@localhost"
 
-if [ "${CONTAINER_RUNTIME}" = "containerd"  ] ; then
-   # $ssh_cmd docker network prune 2>/dev/null
-    $ssh_cmd systemctl stop docker 2>/dev/null
-    $ssh_cmd systemctl disable docker 2>/dev/null
-   # $ssh_cmd ip link delete cni0 2>/dev/null
-    if [ -z "${CONTAINERD_TARBALL_URL}"  ] ; then
-        CONTAINERD_TARBALL_URL="https://github.com/containerd/containerd/releases/download/v${CONTAINERD_VERSION}/cri-containerd-cni-${CONTAINERD_VERSION}-linux-amd64.tar.gz"
-    fi
+is_true() {
+    [ "$(echo "${1:-false}" | tr '[:upper:]' '[:lower:]')" = "true" ]
+}
 
-    $ssh_cmd curl --retry 5 --retry-delay 10 -L ${CONTAINERD_TARBALL_URL} -o /srv/magnum/cri-containerd-cni.tar.gz
-    $ssh_cmd tar xzvf /srv/magnum/cri-containerd-cni.tar.gz -C / --no-same-owner --touch --no-same-permissions --exclude=etc/cni/net.d --exclude=opt/cni/bin --exclude="*.txt" --exclude=opt/containerd/cluster/gce
-    $ssh_cmd mkdir -p /etc/containerd
+if ! is_true "${IS_UPGRADE:-false}" && ! is_true "${IS_RESIZE:-false}"; then
+    echo "END: install cri"
+else
+    if [ "${CONTAINER_RUNTIME}" = "containerd"  ] ; then
+       # $ssh_cmd docker network prune 2>/dev/null
+        $ssh_cmd systemctl stop docker 2>/dev/null
+        $ssh_cmd systemctl disable docker 2>/dev/null
+       # $ssh_cmd ip link delete cni0 2>/dev/null
+        if [ -z "${CONTAINERD_TARBALL_URL}"  ] ; then
+            CONTAINERD_TARBALL_URL="https://github.com/containerd/containerd/releases/download/v${CONTAINERD_VERSION}/cri-containerd-cni-${CONTAINERD_VERSION}-linux-amd64.tar.gz"
+        fi
+
+        $ssh_cmd curl --retry 5 --retry-delay 10 -L ${CONTAINERD_TARBALL_URL} -o /srv/magnum/cri-containerd-cni.tar.gz
+        $ssh_cmd tar xzvf /srv/magnum/cri-containerd-cni.tar.gz -C / --no-same-owner --touch --no-same-permissions --exclude=etc/cni/net.d --exclude=opt/cni/bin --exclude="*.txt" --exclude=opt/containerd/cluster/gce
+        $ssh_cmd mkdir -p /etc/containerd
 cat << EOF > /etc/containerd/config.toml
 version = 2
 root = "/var/lib/containerd"
@@ -59,24 +66,25 @@ oom_score = 0
     path = "/var/lib/containerd/opt"
 EOF
 
-    $ssh_cmd systemctl daemon-reload
-    $ssh_cmd systemctl enable containerd
-    $ssh_cmd systemctl start containerd
-else
-    # CONTAINER_RUNTIME=host-docker
-    $ssh_cmd systemctl disable docker
-    if $ssh_cmd cat /usr/lib/systemd/system/docker.service | grep 'native.cgroupdriver'; then
-            $ssh_cmd cp /usr/lib/systemd/system/docker.service /etc/systemd/system/
-            sed -i "s/\(native.cgroupdriver=\)\w\+/\1$CGROUP_DRIVER/" \
-                    /etc/systemd/system/docker.service
+        $ssh_cmd systemctl daemon-reload
+        $ssh_cmd systemctl enable containerd
+        $ssh_cmd systemctl start containerd
     else
-            cat > /etc/systemd/system/docker.service.d/cgroupdriver.conf << EOF
+        # CONTAINER_RUNTIME=host-docker
+        $ssh_cmd systemctl disable docker
+        if $ssh_cmd cat /usr/lib/systemd/system/docker.service | grep 'native.cgroupdriver'; then
+                $ssh_cmd cp /usr/lib/systemd/system/docker.service /etc/systemd/system/
+                sed -i "s/\(native.cgroupdriver=\)\w\+/\1$CGROUP_DRIVER/" \
+                        /etc/systemd/system/docker.service
+        else
+                cat > /etc/systemd/system/docker.service.d/cgroupdriver.conf << EOF
     ExecStart=---exec-opt native.cgroupdriver=$CGROUP_DRIVER
 EOF
-    fi
+        fi
 
-    $ssh_cmd systemctl daemon-reload
-    $ssh_cmd systemctl enable docker
+        $ssh_cmd systemctl daemon-reload
+        $ssh_cmd systemctl enable docker
+    fi
 fi
 
 echo "END: install cri"
